BUG: ufunc.at segfaults with non-trivial signature
Calling the method `at` of a ufunc with non-trivial signature can lead
to a segfault if all checks on the shape, number of arguments, etc, pass
and the actual computation is attempted, e.g. for the ufunc `matmul`.
Now, using `at` with a non-trivial signature will raise a TypeError
analogously the method `outer`.